+++ pyClarion/__init__.py
@@ -2,13 +2,13 @@ from .base import dimension, feature, chunk, rule, Module, Structure
 from .components import (Repeat, Receptors, Actions, CAM, Shift, 
     BoltzmannSampler, ActionSampler, BottomUp, TopDown, AssociativeRules, 
     ActionRules, BLATracker, Store, GoalStore, Flags, Slots, Gates, DimFilter)
-from . import numdicts as nd
-from .utils import pprint, load, inspect
+from .numdicts import NumDict
+from .utils import pprint, pformat, load, inspect
 
 __all__ = [
     "dimension", "feature", "chunk", "rule", "Module", "Structure",
     "Repeat", "Receptors", "Actions", "CAM", "Shift", "BoltzmannSampler", 
     "ActionSampler", "BottomUp", "TopDown", "AssociativeRules", "ActionRules", 
     "BLATracker", "Store", "GoalStore", "Flags", "Slots", "Gates", "DimFilter", 
-    "nd", "pprint", "load", "inspect"
+    "NumDict", "pprint", "pformat", "load", "inspect"
 ]