+++ pyClarion/utils/__init__.py
@@ -1,6 +1,6 @@
-from . import pprint
+from .pprint import pprint, pformat
 from .load import load
 from . import inspect
     
 
-__all__ = ["pprint", "load", "inspect"]
+__all__ = ["pprint", "pformat", "load", "inspect"]