+++ changelog.md
@@ -12,9 +12,10 @@ This is a backwards incompatible rewrite.
 ## Added
 
 - A new `dimension` symbol.
-- A markup language and parser for easily specifying initial chunks & fixed rules.
-- Visualization and inspection tools.
 - Support for multiple outputs in each process.
+- A new mini language `ccml` and interpreter for easily specifying initial chunks & rules.
+- The `dev` module, collecting tools and resources relevant for developing new components.
+- Visualization and inspection tools.
 
 ## Changed